Fix parsing of VARCHAR values with length smaller than 255.
Quoting from the "The CHAR and VARCHAR Types" section of the manual:

In contrast to CHAR, VARCHAR values are stored as a 1-byte or 2-byte
length prefix plus data. The length prefix indicates the number of
bytes in the value. A column uses one length byte if values require
no more than 255 bytes, two length bytes if values may require more
than 255 bytes.